+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
 ## Unreleased
+
+- Enhance build-projects-manifest action to support custom manifests. (`4ac4277`)
+